deleting anonymous namespace and libraries
